requirements.txt
rich>=13.0
prompt_toolkit>=3.0
requests>=2.28
edge-tts>=6.0
vosk>=0.3.45
textual>=0.47.1
langgraph>=0.0.26
langchain-core>=0.1.30
langchain-openai>=0.0.8
langchain-community>=0.0.27
chromadb>=0.4.24
sentence-transformers>=2.5.1
duckduckgo-search>=5.2.1
python-docx>=1.1.0
pypdf>=4.0.0
python-pptx>=0.6.23
openpyxl>=3.1.0
